installer: module editor
