+++ gui/src/installer/step/mod.rs
@@ -9,7 +9,9 @@ pub use node::{
     DefineNode,
 };
 
-pub use descriptor::{BackupDescriptor, DefineDescriptor, ImportDescriptor, RegisterDescriptor};
+pub use descriptor::{
+    editor::DefineDescriptor, BackupDescriptor, ImportDescriptor, RegisterDescriptor,
+};
 
 pub use backend::{ChooseBackend, ImportRemoteWallet, RemoteBackendLogin};
 pub use mnemonic::{BackupMnemonic, RecoverMnemonic};
